Changed all configuration fetch calls to directly use getters from underlying FileConfiguration instance Changed a lot of field visibilities to private since that is better coding style
