+++ src/com/Jdbye/BukkitIRCd/BukkitIRCdPlayerListener.java
@@ -1,5 +1,6 @@
 package com.Jdbye.BukkitIRCd;
 
+import com.Jdbye.BukkitIRCd.configuration.Config;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -12,8 +13,6 @@ import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.server.ServerCommandEvent;
-
-import com.Jdbye.BukkitIRCd.configuration.*;
 
 /**
  * Handle events for all Player related events
@@ -28,21 +27,21 @@ public class BukkitIRCdPlayerListener implements Listener {
 
 	@EventHandler(priority = EventPriority.MONITOR)
 	public void onPlayerDeath(PlayerDeathEvent event){
-		if (!IRCd.broadcastDeathMessages){
+		if (!Config.isIrcdBroadCastDeathMessages()){
 			return; 
 		}
-		String message = event.getDeathMessage().replace(event.getEntity().getName(),event.getEntity().getName()+IRCd.ingameSuffix);
+		String message = event.getDeathMessage().replace(event.getEntity().getName(),event.getEntity().getName()+Config.getIrcdIngameSuffix());
 		
-		if(!IRCd.colorDeathMessages){
+		if(!Config.isIrcdColorDeathMessagesEnabled()){
 			message = ChatColor.stripColor(message);
 		}
 		else {
 			message = IRCd.convertColors(message,false);
 		}
 		if(IRCd.mode == Modes.INSPIRCD){
-			if (IRCd.linkcompleted)  {
+			if (IRCd.isLinkcompleted())  {
 				
-				IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + IRCd.channelName + " :" + message);
+				IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + Config.getIrcdChannel() + " :" + message);
 				}
 		}
 		else {
@@ -58,7 +57,7 @@ public class BukkitIRCdPlayerListener implements Listener {
 		
 		if (split.length > 1){
 			
-			if (Config.kickCommands.contains(split[0].toLowerCase())){
+			if (Config.getKickCommands().contains(split[0].toLowerCase())){
 			
 				//PlayerKickEvent does not give kicker, so we listen to kick commands instead
 					StringBuilder s = new StringBuilder(300);
@@ -81,12 +80,12 @@ public class BukkitIRCdPlayerListener implements Listener {
 				
 				
 				message = ChatColor.stripColor(ChatColor.translateAlternateColorCodes('&',message));
-				if(IRCd.colorSayMessages){
+				if(Config.isIrcdColorSayMessageEnabled()){
 					message = (char) 3 + "13" + message;
 				}
 				if(IRCd.mode == Modes.INSPIRCD){
-					if (IRCd.linkcompleted)  {
-						IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + IRCd.channelName + " :" + message);
+					if (IRCd.isLinkcompleted())  {
+						IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + Config.getIrcdChannel() + " :" + message);
 						}
 				}else{
 					IRCd.writeAll(message);
@@ -105,37 +104,37 @@ public class BukkitIRCdPlayerListener implements Listener {
             StringBuffer mode = new StringBuffer();
             Player player = event.getPlayer();
             if (player.hasPermission("bukkitircd.mode.owner")){
-            	if (IRCd.debugMode) {
+            	if (Config.isDebugModeEnabled()) {
             		BukkitIRCdPlugin.log.info("Add mode +q for " + player.getName());
             	}
             	mode.append("~");
             	
             }
             if (player.hasPermission("bukkitircd.mode.protect")){
-            	if (IRCd.debugMode) {
+            	if (Config.isDebugModeEnabled()) {
             		BukkitIRCdPlugin.log.info("Add mode +a for " + player.getName());
             	}
             	mode.append("&");
             }
             if (player.hasPermission("bukkitircd.mode.op")){
-            	if (IRCd.debugMode) {
+            	if (Config.isDebugModeEnabled()) {
             		BukkitIRCdPlugin.log.info("Add mode +o for " + player.getName());
             	}
             	mode.append("@");
             }
             if (player.hasPermission("bukkitircd.mode.halfop")){
-            	if (IRCd.debugMode) {
+            	if (Config.isDebugModeEnabled()) {
             		BukkitIRCdPlugin.log.info("Add mode +h for " + player.getName());
             	}
             	mode.append("%");
             }
             if (player.hasPermission("bukkitircd.mode.voice")){
-            	if (IRCd.debugMode) {
+            	if (Config.isDebugModeEnabled()) {
             		BukkitIRCdPlugin.log.info("Add mode +v for " + player.getName());
             	}
             	mode.append("+");
             }
-            if (!IRCd.redundantModes){
+            if (!Config.isIrcdRedundantModes()){
             	mode.delete(1, mode.length()); //Remove all but the mode powerful mode if redundant modes are not allowed
             }
             IRCd.addBukkitUser(mode.toString(),player);
@@ -160,7 +159,7 @@ public class BukkitIRCdPlayerListener implements Listener {
 		else {
 			BukkitPlayer bp;
 			if ((bp = IRCd.getBukkitUserObject(event.getPlayer().getName())) != null) {
-				if (IRCd.linkcompleted) IRCd.println(":" + bp.getUID() + " PRIVMSG " + IRCd.channelName + " :" + IRCd.convertColors(event.getMessage(), false));
+				if (IRCd.isLinkcompleted()) IRCd.println(":" + bp.getUID() + " PRIVMSG " + Config.getIrcdChannel() + " :" + IRCd.convertColors(event.getMessage(), false));
 			}
 		}
 		event.setMessage(IRCd.stripIRCFormatting(event.getMessage()));
@@ -183,14 +182,14 @@ public class BukkitIRCdPlayerListener implements Listener {
 				else {
 					BukkitPlayer bp;
 					if ((bp = IRCd.getBukkitUserObject(event.getPlayer().getName())) != null) {
-						if (IRCd.linkcompleted) IRCd.println(":" + bp.getUID() + " PRIVMSG " + IRCd.channelName + " :" + (char)1 + "ACTION " + IRCd.convertColors(IRCd.join(event.getMessage().split(" ")," ",1), false) + (char)1);
+						if (IRCd.isLinkcompleted()) IRCd.println(":" + bp.getUID() + " PRIVMSG " + Config.getIrcdChannel() + " :" + (char)1 + "ACTION " + IRCd.convertColors(IRCd.join(event.getMessage().split(" ")," ",1), false) + (char)1);
 					}
 				}
 				event.setMessage(IRCd.stripIRCFormatting(event.getMessage()));
 			}
 			
 			
-			if (Config.kickCommands.contains(split[0].substring(1).toLowerCase())){
+			if (Config.getKickCommands().contains(split[0].substring(1).toLowerCase())){
 				//PlayerKickEvent does not give kicker, so we listen to kick commands instead
 				if (event.getPlayer().hasPermission("bukkitircd.kick")){
 					StringBuilder s = new StringBuilder(300);
@@ -220,14 +219,14 @@ public class BukkitIRCdPlayerListener implements Listener {
 					String message = s.toString();
 					
 					message = ChatColor.translateAlternateColorCodes('&',ChatColor.stripColor(message));
-					if(IRCd.colorSayMessages){
+					if(Config.isIrcdColorSayMessageEnabled()){
 						message = (char) 3 + "13" + message;
 					}
 					
 					if(IRCd.mode == Modes.INSPIRCD){
 						
-						if (IRCd.linkcompleted) { 
-							IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + IRCd.channelName + " :" + message);
+						if (IRCd.isLinkcompleted()) {
+							IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + Config.getIrcdChannel() + " :" + message);
 							}
 					}else{
 						IRCd.writeAll(message);
