+++ src/com/Jdbye/BukkitIRCd/IRCCommandSender.java
@@ -1,9 +1,6 @@
 package com.Jdbye.BukkitIRCd;
 
-import java.util.Set;
-import java.util.regex.PatternSyntaxException;
-
-// import org.bukkit.ChatColor;
+import com.Jdbye.BukkitIRCd.configuration.Config;
 import org.bukkit.Server;
 import org.bukkit.command.CommandSender;
 import org.bukkit.permissions.PermissibleBase;
@@ -11,6 +8,11 @@ import org.bukkit.permissions.Permission;
 import org.bukkit.permissions.PermissionAttachment;
 import org.bukkit.permissions.PermissionAttachmentInfo;
 import org.bukkit.plugin.Plugin;
+
+import java.util.Set;
+import java.util.regex.PatternSyntaxException;
+
+// import org.bukkit.ChatColor;
 
 public class IRCCommandSender implements CommandSender {
     private Server server;
@@ -21,7 +23,7 @@ public class IRCCommandSender implements CommandSender {
 	public IRCCommandSender(Server server) {
 		super();
         this.server = server;
-        this.nick = IRCd.serverName;
+        this.nick = Config.getIrcdServerName();
         enabled = true;
     }
 
@@ -40,8 +42,8 @@ public class IRCCommandSender implements CommandSender {
 		}
     	
 
-    			if (IRCd.mode == Modes.STANDALONE) IRCd.writeOpers(":" + IRCd.serverName + "!" + IRCd.serverName + "@" + IRCd.serverHostName + " PRIVMSG " + IRCd.consoleChannelName + " :" + IRCd.convertColors(message, false));
-    			else if (IRCd.linkcompleted) IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + IRCd.consoleChannelName + " :" + IRCd.convertColors(message, false));
+    			if (IRCd.mode == Modes.STANDALONE) IRCd.writeOpers(":" + Config.getIrcdServerName() + "!" + Config.getIrcdServerName() + "@" + Config.getIrcdServerHostName() + " PRIVMSG " + Config.getIrcdConsoleChannel() + " :" + IRCd.convertColors(message, false));
+    			else if (IRCd.isLinkcompleted()) IRCd.println(":" + IRCd.serverUID + " PRIVMSG " + Config.getIrcdConsoleChannel() + " :" + IRCd.convertColors(message, false));
     		}
