+++ src/com/Jdbye/BukkitIRCd/commands/IRCLinkCommand.java
@@ -1,14 +1,14 @@
 package com.Jdbye.BukkitIRCd.commands;
 
+import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
+import com.Jdbye.BukkitIRCd.IRCd;
+import com.Jdbye.BukkitIRCd.Modes;
+import com.Jdbye.BukkitIRCd.configuration.Config;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-
-import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
-import com.Jdbye.BukkitIRCd.IRCd;
-import com.Jdbye.BukkitIRCd.Modes;
 
 public class IRCLinkCommand implements CommandExecutor{
 
@@ -23,13 +23,13 @@ public class IRCLinkCommand implements CommandExecutor{
 			Player player = (Player) sender;
 			if (player.hasPermission("bukkitircd.link")) {
 				if ((IRCd.mode == Modes.INSPIRCD) || (IRCd.mode == Modes.UNREALIRCD)) {
-					if ((!IRCd.linkcompleted) && (!IRCd.isConnected())) {
-						if (IRCd.connect()) player.sendMessage(ChatColor.RED + "Successfully connected to " + IRCd.remoteHost + " on port " + IRCd.remotePort);
-						else player.sendMessage(ChatColor.RED + "Failed to connect to " + IRCd.remoteHost + " on port " + IRCd.remotePort);
+					if ((!IRCd.isLinkcompleted()) && (!IRCd.isConnected())) {
+						if (IRCd.connect()) player.sendMessage(ChatColor.RED + "Successfully connected to " + Config.getLinkRemoteHost() + " on port " + Config.getLinkRemotePort());
+						else player.sendMessage(ChatColor.RED + "Failed to connect to " + Config.getLinkRemoteHost() + " on port " + Config.getLinkRemotePort());
 					}
 					else {
-						if (IRCd.linkcompleted) player.sendMessage(ChatColor.RED + "Already linked to " + IRCd.linkName + ".");
-						else player.sendMessage(ChatColor.RED + "Already connected to " + IRCd.linkName + ", but not linked.");
+						if (IRCd.isLinkcompleted()) player.sendMessage(ChatColor.RED + "Already linked to " + Config.getLinkName() + ".");
+						else player.sendMessage(ChatColor.RED + "Already connected to " + Config.getLinkName() + ", but not linked.");
 					}
 				}
 				else { player.sendMessage(ChatColor.RED + "[BukkitIRCd] You are currently in standalone mode. To link to a server, modify the config."); }
@@ -41,13 +41,13 @@ public class IRCLinkCommand implements CommandExecutor{
 		}else{
 			
 			if ((IRCd.mode == Modes.INSPIRCD) || (IRCd.mode == Modes.UNREALIRCD)) {
-				if ((!IRCd.linkcompleted) && (!IRCd.isConnected())) {
-					if (IRCd.connect()) sender.sendMessage(ChatColor.RED + "Successfully connected to " + IRCd.remoteHost + " on port " + IRCd.remotePort);
-					else sender.sendMessage(ChatColor.RED + "Failed to connect to " + IRCd.remoteHost + " on port " + IRCd.remotePort);
+				if ((!IRCd.isLinkcompleted()) && (!IRCd.isConnected())) {
+					if (IRCd.connect()) sender.sendMessage(ChatColor.RED + "Successfully connected to " + Config.getLinkRemoteHost() + " on port " + Config.getLinkRemotePort());
+					else sender.sendMessage(ChatColor.RED + "Failed to connect to " + Config.getLinkRemoteHost() + " on port " + Config.getLinkRemotePort());
 				}
 				else {
-					if (IRCd.linkcompleted) sender.sendMessage(ChatColor.RED + "Already linked to " + IRCd.linkName + ".");
-					else sender.sendMessage(ChatColor.RED + "Already connected to " + IRCd.linkName + ", but not linked.");
+					if (IRCd.isLinkcompleted()) sender.sendMessage(ChatColor.RED + "Already linked to " + Config.getLinkName() + ".");
+					else sender.sendMessage(ChatColor.RED + "Already connected to " + Config.getLinkName() + ", but not linked.");
 				}
 			}
 			else { sender.sendMessage(ChatColor.RED + "[BukkitIRCd] You are currently in standalone mode. To link to a server, modify the config."); }
