+++ src/com/Jdbye/BukkitIRCd/commands/IRCMsgCommand.java
@@ -1,16 +1,16 @@
 package com.Jdbye.BukkitIRCd.commands;
 
-import org.bukkit.ChatColor;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-
 import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
 import com.Jdbye.BukkitIRCd.BukkitPlayer;
 import com.Jdbye.BukkitIRCd.IRCUser;
 import com.Jdbye.BukkitIRCd.IRCd;
 import com.Jdbye.BukkitIRCd.Modes;
+import com.Jdbye.BukkitIRCd.configuration.Config;
+import org.bukkit.ChatColor;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 
 public class IRCMsgCommand implements CommandExecutor {
 
@@ -33,7 +33,7 @@ public class IRCMsgCommand implements CommandExecutor {
 									ircuser.nick,
 									":"
 											+ player.getName()
-											+ IRCd.ingameSuffix
+											+ Config.getIrcdIngameSuffix()
 											+ "!"
 											+ player.getName()
 											+ "@"
@@ -68,7 +68,7 @@ public class IRCMsgCommand implements CommandExecutor {
 									.getBukkitUserObject(player.getName())) != null) {
 								String UID = IRCd.getUIDFromIRCUser(ircuser);
 								if (UID != null) {
-									if (IRCd.linkcompleted) {
+									if (IRCd.isLinkcompleted()) {
 										IRCd.println(":"
 												+ bp.getUID()
 												+ " PRIVMSG "
@@ -136,11 +136,11 @@ public class IRCMsgCommand implements CommandExecutor {
 						IRCd.writeTo(
 								ircuser.nick,
 								":"
-										+ IRCd.serverName
+										+ Config.getIrcdServerName()
 										+ "!"
-										+ IRCd.serverName
+										+ Config.getIrcdServerName()
 										+ "@"
-										+ IRCd.serverHostName
+										+ Config.getIrcdServerHostName()
 										+ " PRIVMSG "
 										+ ircuser.nick
 										+ " :"
@@ -164,7 +164,7 @@ public class IRCMsgCommand implements CommandExecutor {
 					} else if (IRCd.mode == Modes.INSPIRCD) {
 						String UID = IRCd.getUIDFromIRCUser(ircuser);
 						if (UID != null) {
-							if (IRCd.linkcompleted) {
+							if (IRCd.isLinkcompleted()) {
 								IRCd.println(":"
 										+ IRCd.serverUID
 										+ " PRIVMSG "
