+++ src/com/Jdbye/BukkitIRCd/commands/IRCReplyCommand.java
@@ -1,16 +1,16 @@
 package com.Jdbye.BukkitIRCd.commands;
 
-import org.bukkit.ChatColor;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-
 import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
 import com.Jdbye.BukkitIRCd.BukkitPlayer;
 import com.Jdbye.BukkitIRCd.IRCUser;
 import com.Jdbye.BukkitIRCd.IRCd;
 import com.Jdbye.BukkitIRCd.Modes;
+import com.Jdbye.BukkitIRCd.configuration.Config;
+import org.bukkit.ChatColor;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 
 public class IRCReplyCommand implements CommandExecutor {
 
@@ -39,7 +39,7 @@ public class IRCReplyCommand implements CommandExecutor {
 										ircuser.nick,
 										":"
 												+ player.getName()
-												+ IRCd.ingameSuffix
+												+ Config.getIrcdIngameSuffix()
 												+ "!"
 												+ player.getName()
 												+ "@"
@@ -74,7 +74,7 @@ public class IRCReplyCommand implements CommandExecutor {
 									String UID = IRCd
 											.getUIDFromIRCUser(ircuser);
 									if (UID != null) {
-										if (IRCd.linkcompleted) {
+										if (IRCd.isLinkcompleted()) {
 											IRCd.println(":"
 													+ bp.getUID()
 													+ " PRIVMSG "
@@ -154,11 +154,11 @@ public class IRCReplyCommand implements CommandExecutor {
 							IRCd.writeTo(
 									ircuser.nick,
 									":"
-											+ IRCd.serverName
+											+ Config.getIrcdServerName()
 											+ "!"
-											+ IRCd.serverName
+											+ Config.getIrcdServerName()
 											+ "@"
-											+ IRCd.serverHostName
+											+ Config.getIrcdServerHostName()
 											+ " PRIVMSG "
 											+ ircuser.nick
 											+ " :"
@@ -184,7 +184,7 @@ public class IRCReplyCommand implements CommandExecutor {
 						} else if (IRCd.mode == Modes.INSPIRCD) {
 							String UID = IRCd.getUIDFromIRCUser(ircuser);
 							if (UID != null) {
-								if (IRCd.linkcompleted) {
+								if (IRCd.isLinkcompleted()) {
 									IRCd.println(":"
 											+ IRCd.serverUID
 											+ " PRIVMSG "
