+++ src/com/Jdbye/BukkitIRCd/commands/IRCTopicCommand.java
@@ -1,13 +1,13 @@
 package com.Jdbye.BukkitIRCd.commands;
 
+import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
+import com.Jdbye.BukkitIRCd.IRCd;
+import com.Jdbye.BukkitIRCd.configuration.Config;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-
-import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
-import com.Jdbye.BukkitIRCd.IRCd;
 
 public class IRCTopicCommand implements CommandExecutor{
 
@@ -25,7 +25,7 @@ public class IRCTopicCommand implements CommandExecutor{
 					String topic = IRCd.join(args, " ", 0);
 					String playername = player.getName();
 					String playerhost = player.getAddress().getAddress().getHostName();
-					IRCd.setTopic(IRCd.convertColors(topic, false), playername + IRCd.ingameSuffix, playername + IRCd.ingameSuffix + "!" + playername + "@" + playerhost);
+					IRCd.setTopic(IRCd.convertColors(topic, false), playername + Config.getIrcdIngameSuffix(), playername + Config.getIrcdIngameSuffix() + "!" + playername + "@" + playerhost);
 					player.sendMessage(ChatColor.RED + "Topic set to " + topic);
 				}
 				else { player.sendMessage(ChatColor.RED + "Please provide a topic to set."); return false; }
@@ -37,7 +37,7 @@ public class IRCTopicCommand implements CommandExecutor{
 		}else{
 			if (args.length > 0) {
 				String topic = IRCd.join(args, " ", 0);
-				IRCd.setTopic(IRCd.convertColors(topic, false), IRCd.serverName, IRCd.serverName + "!" + IRCd.serverName + "@" + IRCd.serverHostName);
+				IRCd.setTopic(IRCd.convertColors(topic, false), Config.getIrcdServerName(), Config.getIrcdServerName() + "!" + Config.getIrcdServerName() + "@" + Config.getIrcdServerHostName());
 				sender.sendMessage(ChatColor.RED + "Topic set to " + topic);
 			}
 			else { sender.sendMessage(ChatColor.RED + "Please provide a topic to set."); return false; }
