+++ src/com/Jdbye/BukkitIRCd/commands/RawsendCommand.java
@@ -1,15 +1,14 @@
 package com.Jdbye.BukkitIRCd.commands;
 
+import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
+import com.Jdbye.BukkitIRCd.IRCd;
+import com.Jdbye.BukkitIRCd.Modes;
+import com.Jdbye.BukkitIRCd.configuration.Config;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-
-import com.Jdbye.BukkitIRCd.BukkitIRCdPlugin;
-import com.Jdbye.BukkitIRCd.IRCd;
-import com.Jdbye.BukkitIRCd.Modes;
-import com.Jdbye.BukkitIRCd.configuration.*;
 
 public class RawsendCommand implements CommandExecutor{
 
@@ -21,15 +20,14 @@ public class RawsendCommand implements CommandExecutor{
 	public boolean onCommand(CommandSender sender, Command cmd, String label,
 			String[] args) {
 		if (sender instanceof Player){
-			sender = (Player) sender;
-			if (Config.enableRawSend) {
+			if (Config.isEnableRawSend()) {
 				sender.sendMessage(ChatColor.RED + "[BukkitIRCd] Only the console can use this command.");
 			}
 			else { sender.sendMessage(ChatColor.RED + "[BukkitIRCd] Sending raw messages is disabled. Please enable them in the config first."); }
 			return true;
 		
 		}else{
-			if (Config.enableRawSend) {
+			if (Config.isEnableRawSend()) {
 				if (args.length > 0) {
 					if ((IRCd.mode == Modes.INSPIRCD) || (IRCd.mode == Modes.UNREALIRCD)) {
 						if (IRCd.println(IRCd.join(args, " ", 0))) sender.sendMessage(ChatColor.RED + "Command sent to IRC server link.");
